[quant][graphmode][fx] Optimize cat
Summary:
Previously we have a cat that takes a list of Tensors with different qparams and dequantize them
cacatenate them and requantize with the output qparams. This adds some unnecessary overhead in dequantizing
and quantizing Tensors.

This PR adds an optimization for cat operator, we'll make sure inputs and output of cat
uses same observer/fake_quant and produce a cat that does not do rescaling.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: f1d548490307fbc094146ab60fe6cfe6fd16290b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54813